Combine generator with base class
