--- generators/dotnetcore-server-interfaces/PrincipleStudios.OpenApiCodegen.Server.Mvc/OpenApiGeneratorOptions.cs
@@ -1,6 +1,0 @@
-﻿using Microsoft.OpenApi.Models;
-
-namespace PrincipleStudios.OpenApiCodegen.Server.Mvc
-{
-    public record OpenApiGeneratorOptions(OpenApiDocument Document);
-}
