--- generators/dotnetstandard-client/PrincipleStudios.OpenApiCodegen.Client/OpenApiGeneratorOptions.cs
@@ -1,6 +1,0 @@
-﻿using Microsoft.OpenApi.Models;
-
-namespace PrincipleStudios.OpenApiCodegen.Client
-{
-    public record OpenApiGeneratorOptions(OpenApiDocument Document);
-}
